requirements.txt
2captcha-python==1.2.2
alabaster==1.0.0
annotated-types==0.5.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
bertopic==0.16.2
bibtexparser==1.4.3
bleach==6.1.0
blinker==1.6.2
blis==0.7.10
BTM==1.1.5
catalogue==2.0.9
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.10
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.2.0
confection==0.1.1
contourpy==1.3.0
cryptography==41.0.3
cycler==0.12.1
cymem==2.0.7
Cython==0.29.37
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.7
docutils==0.21.2
EasyProcess==1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0-py3-none-any.whl#sha256=83276fc78a70045627144786b52e1f2728ad5e29e5e43916ec37ea9c26a11212
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.1
faiss-cpu==1.7.4
fake-useragent==2.2.0
fastapi==0.99.1
fastjsonschema==2.20.0
filelock==3.12.2
Flask==2.3.3
Flask-Cors==4.0.0
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
flatbuffers==23.5.26
fonttools==4.53.1
fqdn==1.5.1
free_proxy==1.1.3
fsspec==2023.9.0
funcy==2.0
gensim==4.3.2
graphql-core==3.2.3
h11==0.14.0
hdbscan==0.8.37
httpcore==1.0.5
httptools==0.6.0
httpx==0.27.2
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imagesize==1.4.1
importlib-resources==6.0.1
ipykernel==6.29.4
ipython==8.26.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
jwt==1.3.1
keyboard==0.13.5
kiwisolver==1.4.7
langcodes==3.3.0
langdetect==1.0.9
llvmlite==0.43.0
lxml==5.4.0
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.7
mistune==3.0.2
monotonic==1.6
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
murmurhash==1.0.9
mysql-connector-python==9.3.0
mysqlclient==2.2.7
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.1
nltk==3.8.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numexpr==2.10.1
numpy==1.25.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cudnn-cu11==8.9.4.25
onnxruntime==1.15.1
opencv-python==4.8.1.78
openpyxl==3.1.2
outcome==1.2.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
pathy==0.10.2
Pillow==10.0.0
platformdirs==3.10.0
plotly==5.22.0
posthog==3.0.2
preshed==3.0.8
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.24.3
psutil==6.0.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pulsar-client==3.3.0
pure-eval==0.2.2
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.6.3
pygame==2.5.2
PyGetWindow==0.0.9
Pygments==2.18.0
PyJWT==2.8.0
pyLDAvis==3.4.1
PyMsgBox==1.0.9
PyMySQL==1.1.1
pynndescent==0.5.13
pyparsing==3.1.4
PyPDF2==3.0.1
pyperclip==1.8.2
PyPika==0.48.9
pyreadline3==3.4.1
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytweening==1.0.7
pytz==2023.3
pywin32==306
pywinauto==0.6.8
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2023.8.8
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.20.0
safetensors==0.3.3
Sastrawi==1.0.1
scapy==2.5.0
scholarly==1.7.11
scikit-fuzzy==0.5.0
scikit-learn==1.3.0
scipy==1.11.2
selenium==4.11.2
Send2Trash==1.8.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
srsly==2.4.7
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tenacity==8.4.2
terminado==0.18.1
tesseract==0.1.3
thinc==8.1.12
threadpoolctl==3.2.0
tinycss2==1.3.0
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
transformers==4.32.1
trio==0.22.2
trio-websocket==0.10.3
txtai==6.0.0
typer==0.9.0
types-python-dateutil==2.9.0.20240906
typing_extensions==4.7.1
tzdata==2023.3
umap-learn==0.5.6
uri-template==1.3.0
urllib3==2.0.4
uvicorn==0.23.2
virtualenv==20.24.3
wasabi==1.1.2
watchfiles==0.20.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==2.3.7
widgetsnbextension==4.0.13
wrapt==1.17.2
wsproto==1.2.0
